Add parent class to 2 parsers for sort
